+++ tianshou/utils/__init__.py
@@ -1,6 +1,7 @@
 from tianshou.utils.config import tqdm_config
 from tianshou.utils.statistics import MovAvg, RunningMeanStd
 from tianshou.utils.log_tools import BasicLogger, LazyLogger, BaseLogger
+from tianshou.utils.optimizer import KFACOptimizer
 
 __all__ = [
     "MovAvg",
@@ -9,4 +10,5 @@ __all__ = [
     "BaseLogger",
     "BasicLogger",
     "LazyLogger",
+    "KFACOptimizer",
 ]
